deployment/requirements.txt
aws-cdk-lib==2.212.0
constructs>=10.0.0,<11.0.0
cdk_nag==2.37.8